backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-socketio==5.11.2
python-multipart==0.0.9
httpx==0.27.0
aiohttp==3.9.5
websockets==12.0

# ── LLM / AI ──────────────────────────────────────────────────────────────
google-generativeai==0.7.2
langchain==0.2.5
langchain-community==0.2.5
langchain-google-genai==1.0.6
openai==1.35.3

# ── Vector DB / Memory ─────────────────────────────────────────────────────
faiss-cpu==1.8.0
chromadb==0.5.3
sentence-transformers==3.0.1

# ── Knowledge Graph ────────────────────────────────────────────────────────
neo4j==5.21.0

# ── NLP ───────────────────────────────────────────────────────────────────
textblob==0.18.0
nltk==3.8.1
spacy==3.7.4

# ── Vision / OCR ──────────────────────────────────────────────────────────
pytesseract==0.3.10
Pillow==10.3.0
opencv-python-headless==4.10.0.82

# ── System / OS ───────────────────────────────────────────────────────────
psutil==6.0.0
pyautogui==0.9.54
pywin32==306; sys_platform == "win32"
pynput==1.7.6
watchdog==4.0.0

# ── Data / Compute ────────────────────────────────────────────────────────
numpy==1.26.4
scipy==1.13.1
networkx==3.3
pandas==2.2.2
joblib==1.4.2

# ── Scraping / OSINT ──────────────────────────────────────────────────────
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.44.0
selenium==4.21.0
feedparser==6.0.10

# ── Security ──────────────────────────────────────────────────────────────
cryptography==42.0.8
scapy==2.5.0

# ── Async / Concurrency ───────────────────────────────────────────────────
asyncio==3.4.3
aiofiles==23.2.1
aiocache==0.12.2

# ── Media / Video ─────────────────────────────────────────────────────────
moviepy==1.0.3
pydub==0.25.1
SpeechRecognition==3.10.4
yt-dlp==2024.4.9

# ── Storage ───────────────────────────────────────────────────────────────
sqlalchemy==2.0.31
alembic==1.13.1
redis==5.0.6

# ── Utils ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.4
loguru==0.7.2
rich==13.7.1
typer==0.12.3
schedule==1.2.2
setuptools
wheel
